Updated the build.gradle.kts file to be able to run UnitTest in my customer-service.
Created a unitTest for my customerCustomerController

Modified the ports for the mysql database and the name of the server
for all services inside the compose.yml file to be able to differentiate
them  and matched the ports in the application.yml file.

Created the Api-Gateway module and modified the compose.yaml. Removed all PhpMyAdmin container creation from other services and only put it inside the Api-Gateway's compose.yaml as we only want tu use one PhPMyAdmin container to access the other mysql servers.